+++ custom_components/worx_vision_cloud/const.py
@@ -22,6 +22,7 @@ PLATFORMS = [
     Platform.DEVICE_TRACKER,
     Platform.SENSOR,
     Platform.BINARY_SENSOR,
+    Platform.SWITCH,
     Platform.BUTTON,
 ]